Update files and preprocessing
Updated the files that preprocess the text to split text into sentences.
moved all the old code to OldCode.py
